pyproject.toml
[tool.poetry]
name = "oop-2026"
version = "0.1.0"
description = ""
authors = ["Piotr Marecki <pmarecki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.12.0"
pillow = "^12.0.0"
python-dotenv = "^1.2.1"
openai = "^2.14.0"
xai-sdk = "^1.5.0"
google-genai = "^1.56.0"
loguru = "^0.7.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
